fix: User 필드 타입 수정
userCode 타입 int -> Long
loginType, userLevel Enum으로 수정

+++ src/main/java/com/life/muna/user/dto/SignInResponse.java
@@ -6,12 +6,13 @@ import lombok.Getter;
 
 @Getter
 public class SignInResponse {
-    private final int userCode;
+    private final Long userCode;
     private final TokenResponse token;
 
     @Builder
-    public SignInResponse(int userCode, TokenResponse tokenResponse) {
+    public SignInResponse(Long userCode, TokenResponse tokenResponse) {
         this.userCode = userCode;
         this.token = tokenResponse;
     }
+
 }
